requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
apache-beam==2.58.1
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
botocore==1.35.5
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
comm==0.2.2
crcmod==1.7
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
dill==0.3.1.1
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
executing==2.0.1
fastavro==1.9.7
fasteners==0.19
frozenlist==1.4.1
fsspec==2023.12.2
gcsfs==2023.12.2.post1
google-api-core==2.19.0
google-apitools==0.5.31
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.65.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.26.0
google-cloud-bigtable==2.26.0
google-cloud-core==2.4.1
google-cloud-datastore==2.20.1
google-cloud-dlp==3.22.0
google-cloud-language==2.14.0
google-cloud-pubsub==2.23.1
google-cloud-pubsublite==1.11.1
google-cloud-recommendations-ai==0.10.12
google-cloud-resource-manager==1.12.5
google-cloud-spanner==3.49.1
google-cloud-storage==2.17.0
google-cloud-videointelligence==2.13.5
google-cloud-vision==3.7.4
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
hdfs==2.7.3
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jmespath==1.0.1
Js2Py==0.74
jsonpatch==1.33
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.7
langchain-experimental==0.0.61
langchain-text-splitters==0.2.1
langsmith==0.1.77
markdown-it-py==3.0.0
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
mmhash3==3.0.1
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.7.0
openai==1.34.0
orjson==3.10.5
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pydot==1.4.2
Pygments==2.18.0
pyiceberg==0.7.1
pyjsparser==2.7.1
pymetastore==0.4.1
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
sqlparse==0.5.1
stack-data==0.6.3
strictyaml==1.7.3
tenacity==8.4.1
thrift==0.20.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
wcwidth==0.2.13
yarl==1.9.4
zstandard==0.23.0
